Solve part 2!
I needed to run a check if the box could be moved before actually
moving. The bug was appearing when trying to move a box vertically, and
that box branches into two boxes: if one of them cannot be moved, either
of them should be moved. Therefore I needed to check if both boxes could
be moved before actually performing the move.