Solve day 20 part 2
Write a new implementation for the two parts that avoid walking the map
looking for cheats, but instead gather the points at a given Manhattan
distance for a current point and check the time difference between it
and the other ones. If the time difference minus the Manhattan distance
is positive, then we can use a cheat and save time. Limit the Manhattan
distance to cover up to the maximum cheat time.